Translated scaled_exp1
