+++ scipy/special/functions.json
@@ -493,11 +493,6 @@
             "eval_sh_legendre_l": "ld->d"
         }
     },
-    "_scaled_exp1": {
-        "scaled_exp1.h": {
-            "scaled_exp1": "d->d"
-        }
-    },
     "exp10": {
         "cephes.h": {
             "exp10": "d->d"
