--- scipy/special/scaled_exp1.h
@@ -1,6 +1,0 @@
-#ifndef SCALED_EXP1_H
-#define SCALED_EXP1_H
-
-extern double scaled_exp1(double);
-
-#endif
